requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
daphne==4.0.0
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
gtfs-realtime-bindings==0.0.7
gunicorn==20.1.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
markdown2==2.4.3
msgpack==1.0.4
packaging==21.3
Pillow==9.2.0
priority==1.3.0
protobuf==4.21.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytz==2022.4
redis==4.3.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==22.10.0
txaio==22.2.1
typing-extensions==4.4.0
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.14.1
zope.interface==5.5.0
